Fix missing critical path components due to change-pruning.
The `CritialPathComputer` listens action execution related events to re-construct the action graph. And compute the critical path out of it.

However, during an incremental build, actions might be change-pruned if all of their inputs are not changed after re-execution. It means the action execution events will not be emitted for these actions. The `CriticalPathComputer` will only see a partial action graph resulting in wrong critical path.

This CL fixes that by querying skyframe for the missing generating action in the case of change-pruning.

PiperOrigin-RevId: 691464331
Change-Id: I70c9afbc56ff8c5375cfeff01071be2d6ed95494