Add custom `Predicate<SkyKey> keepGoing` test coverage in `ParallelEvaluatorTest`.
Demonstrate the custom predicate by asserting on the number of evaluated keys. A parent key with an error dep is not evaluated if the dep is evaluated with `--nokeep_going`. On the other hand, a parent key with an error dep is evaluated if the dep is evaluated with `--keep_going`.

PiperOrigin-RevId: 692223929
Change-Id: Id67244d9da7a241a8437fa89b8a517de526e0bf9