Handle action-less RemoteConfiguredTargetValues in CriticalPathComputer.
The critical path computation was recently reworked to query skyframe for the
owning analysis node of an action that produced an input artifact to another
action in some cases, but in analysis caching builds, the analysis node may be
a deserialized analysis node and doesn't contain actions. For these cases, skip
the action for critical path computation.

PiperOrigin-RevId: 693363716
Change-Id: Iace4e723274807d5e69483b9841ac2f0aa0029b4