Stop exporting proto_common and ProtoInfo
This basically removes ProtoInfo and proto_common from Bazel. Keeping the code for internal purposes.

Set --incompatible_autoload_externally to @protobuf (no more providers for @protobuf) and set Java rules that were already removed to a stricter setting (no plus) and provider to +. This is the consistency requirement that native ProtoInfo is not used in anywhere.

Strip down proto_common_do_not_use to bare minimum. The INCOMPATIBLE_ENABLE_TOOLCHAIN_RESOLUTION is needed, because it's a native build language flag, and this is the only way Protobuf can read it.

Fix providers to point to right paths. They are mostly used in tests. ProtoInfo is also used in cc_shared_library, where we use both built-in and Protobuf path, to smooth the migration.

Remove empty `--incompatible_autoload_externally=` from some tests. They just work now without the flag.

PiperOrigin-RevId: 695261259
Change-Id: I677efb8ed94bf6cae8f4c99dd86f657620057850