Remove CONFIGURED_TARGET dependency from TOP_LEVEL_ASPECTS
Get `ConfiguredTargetKey` in `ToplevelStarlarkAspectFunction` without waiting for the full configuration of top level target. This means that the configuration of the aspects (`AspectFuntion`) could start before top level target finishes its configuration.

Before this change, waiting for the top level target to be configured was necessary (which basically means waiting for the configuration of all deps of the top level target).

Moreover, setting `setShouldApplyRuleTransition` to false in the newly created configured key is needed for non-idempotent rule transitions (otherwise, there would be failures, mostly in Bazel).

Furthermore, test is added inside `StarlarkDefinedAspectsTest.java` for this change.

Note that this is the first step towards implementing parallel aspects - Starting aspects ASAP (as soon as we have at least one dependency configured).

PiperOrigin-RevId: 696170160
Change-Id: I9dbb9f07c405f5293142674217a0cc7f5cf7dd75