Remove `GoogleTestSecurityManager` from `BugReportTest`.
When `GoogleTestSecurityManager` is removed, the test's own `ExitProhibitingSecurityManager` is used instead. That enables several additional assertions, and it turns out that those assertions were not correct. This change updates the assertions to match what the code does. Whether that is what the code _should_ do is beyond my current understanding.

PiperOrigin-RevId: 697727042
Change-Id: I235f06b1c4206831553cb2e8fd8d2176387bba37