Load Python providers in unit tests and bazel_tools
Preparing for rules_python upgrade to 0.40.0 and removal of Python providers from Bazel.

Add missing loads for PyInfo and PyRuntimeInfo in tests and in @bazel_tools.
Move PyRuntimeInfo into the testing code. It's not used in production code anymore.
Put PyInfo and PyRuntimeInfo symbols into TestConstants. To support split of  locations internally and in Bazel.

PiperOrigin-RevId: 697976776
Change-Id: Iac93cd927432507249f166e9bf71a422cebed666